Arrays And Lists Assignment
Create a console app that creates a one dimensional array of strings and a one dimensional array of integers and display the value after user selects an index. Create a list of strings and display value after user selects an index. Include a message that displays when user selects an index that doesn't exist.