Classes And Methods Assignment
Create a class with three methods. Each method should take in an integet and execute a math operation on the received parameter. Put this class in a separate .cs file in the app.  Ask user in Main() program for a number and call each method in turn passing the user input to the method.